Refactor invitation logic into InvitationHandler
