+++ frontend/src/components/chat/chatBox/Msg.module.css
@@ -26,5 +26,5 @@
     height: var(--size-picture-profile);
     border-radius: 50%;
     transform: scale(60%);
-	object-fit: cover;
+    object-fit: cover;
 }
